maint: dont require polychord
