+++ src/yabf/samplers/__init__.py
@@ -1,0 +1,11 @@
+"""Implementations of popular samplers."""
+
+try:
+    from .emcee import emcee
+except ImportError:
+    pass
+
+try:
+    from .polychord import polychord
+except ImportError:
+    pass
